abstract class Data
